Consolidated code.
Many separate files were hard to refactor.
This change is purely copy/pasting code into other places and fixing their paths.
