+++ registry/memcache/memcached.go
@@ -1,4 +1,4 @@
-package registry
+package memcache
 
 import (
 	"fmt"
