+++ registry/warming.go
@@ -6,9 +6,11 @@ import (
 	"sync"
 	"time"
 
-	"github.com/bradfitz/gomemcache/memcache"
+	officialMemcache "github.com/bradfitz/gomemcache/memcache"
 	"github.com/go-kit/kit/log"
 	"github.com/pkg/errors"
+	"github.com/weaveworks/flux/registry/memcache"
+	"math/rand"
 	"strings"
 )
 
@@ -17,7 +19,7 @@ type Warmer struct {
 	ClientFactory ClientFactory
 	Creds         Credentials
 	Expiry        time.Duration
-	Client        MemcacheClient
+	Client        memcache.MemcacheClient
 }
 
 // Continuously wait for a new repository to warm
@@ -69,7 +71,7 @@ func (w *Warmer) warm(repository Repository) {
 	// might be duplicates from other registries
 	key := tagKey(username, repository.String())
 
-	if err := w.Client.Set(&memcache.Item{
+	if err := w.Client.Set(&officialMemcache.Item{
 		Key:        key,
 		Value:      val,
 		Expiration: int32(w.Expiry.Seconds()),
@@ -91,7 +93,7 @@ func (w *Warmer) warm(repository Repository) {
 
 		// Full path to image again.
 		key := manifestKey(username, repository.String(), tag)
-		if err := w.Client.Set(&memcache.Item{
+		if err := w.Client.Set(&officialMemcache.Item{
 			Key:        key,
 			Value:      val,
 			Expiration: int32(w.Expiry.Seconds()),
@@ -100,4 +102,58 @@ func (w *Warmer) warm(repository Repository) {
 			return
 		}
 	}
+}
+
+// Queue provides an updating repository queue for the warmer.
+// If no items are added to the queue this will randomly add a new
+// registry to warm
+type Queue struct {
+	RunningContainers    func() []Repository
+	Logger               log.Logger
+	RegistryPollInterval time.Duration
+	warmQueue            chan Repository
+	queueLock            sync.Mutex
+}
+
+func NewQueue(runningContainersFunc func() []Repository, l log.Logger, emptyQueueTick time.Duration) Queue {
+	return Queue{
+		RunningContainers:    runningContainersFunc,
+		Logger:               l,
+		RegistryPollInterval: emptyQueueTick,
+		warmQueue:            make(chan Repository, 100), // Don't close this. It will be GC'ed when this instance is destroyed.
+	}
+}
+
+// Queue loop to maintain the queue and periodically add a random
+// repository that is running in the cluster.
+func (w *Queue) Loop(stop chan struct{}, wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	if w.RunningContainers == nil || w.Logger == nil || w.RegistryPollInterval == 0 {
+		panic("registry.Queue fields are nil")
+	}
+
+	pollImages := time.Tick(w.RegistryPollInterval)
+
+	for {
+		select {
+		case <-stop:
+			w.Logger.Log("stopping", "true")
+			return
+		case <-pollImages:
+			c := w.RunningContainers()
+			if len(c) > 0 { // Only add random registry if there are running containers
+				i := rand.Intn(len(c)) // Pick random registry
+				w.queueLock.Lock()
+				w.warmQueue <- c[i] // Add registry to queue
+				w.queueLock.Unlock()
+			}
+		}
+	}
+}
+
+func (w *Queue) Queue() chan Repository {
+	w.queueLock.Lock()
+	defer w.queueLock.Unlock()
+	return w.warmQueue
 }
